Experiment with components out of routes
Persists a component between pages. In this change, the AppBar is
rendered outside of the top-level router Switch, so that it continues to
animate smoothly (isn't re-rendered) as the user navigates to other
routes in the application.

This loosens the concept of "routes" or "pages" as compositions that are
responsible for the full, visible tree of components. It moves one step
closer to merging "src/routes" into "src/components", and treating
everything exactly the same.